pyproject.toml
[project]
name = "yt-to-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "langgraph>=0.4.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
    "youtube-transcript-api>=1.1.0",
]
